don't parse resource version
Parsing the resourceVersion as an integer is strictly prohibited by the
Kubernetes API [1]. During a review of code that would be affected by
non-numerical resourceVersions, I came upon this code. As the parsed
data is not used for anything, it looks like it's possible to simply
remove this parsing logic.

[1] https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>